requirements.txt
# KathaPe Customer - Production Requirements
# Install these packages for production deployment

# Core Flask
Flask==2.3.3
Werkzeug==2.3.7

# Appwrite NoSQL Database (replaces PostgreSQL)
appwrite==11.1.0

# Legacy PostgreSQL (not needed with Appwrite)
# psycopg2-binary==2.9.7

# Image processing
Pillow==10.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Date/time handling
pytz==2023.3

# WSGI Server
gunicorn==21.2.0

# UUID support (usually built-in)
uuid==1.30

# JSON handling (usually built-in but explicit)
simplejson==3.19.1
